+++ tests/nimble/packageinfo/test_getinstallfiles_binary.nim
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: 2022-2023 Anna <cyber@sysrq.in>
+# SPDX-FileCopyrightText: 2022-2025 Anna <cyber@sysrq.in>
 # SPDX-License-Identifier: BSD-3-Clause
 
 import std/[algorithm, os]
@@ -6,18 +6,18 @@ import nimbs/[options, packageinfo]
 
 const sourceDir = "tests" / "nimble" / "packageinfo" / "binary"
 
-let
-  opts = Options(sourceDir: sourceDir)
-  pkgInfo = PackageInfo(
-    nimbleFile: sourceDir / "binary.nimble",
-    name: "binary",
-    version: "1.0",
-    bin: @["tools/main"],
-    skipFiles: @["README.md"]
-  )
+let opts = Options(sourceDir: sourceDir)
+var pkgInfo = PackageInfo(
+  nimbleFile: sourceDir / "binary.nimble",
+  name: "binary",
+  version: "1.0",
+  bin: @["tools/main"],
+  skipFiles: @["README.md"]
+)
+
+inferInstallRules(pkgInfo, opts)
 
 let files = pkgInfo.getInstallFiles(opts).sorted
 assert files == @[
-  (pcFile, sourceDir / "tools" / "main.nim"),
   (pcDir, sourceDir / "tools")
 ]