+++ tests/nimble/packageinfo/test_getinstallfiles_source.nim
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: 2022-2023 Anna <cyber@sysrq.in>
+# SPDX-FileCopyrightText: 2022-2025 Anna <cyber@sysrq.in>
 # SPDX-License-Identifier: BSD-3-Clause
 
 import std/[algorithm, os]
@@ -6,20 +6,20 @@ import nimbs/[options, packageinfo]
 
 const sourceDir = "tests" / "nimble" / "packageinfo" / "source"
 
-let
-  opts = Options(sourceDir: sourceDir)
-  pkgInfo = PackageInfo(
-    nimbleFile: sourceDir / "source.nimble",
-    name: "source",
-    version: "1.0",
-    skipDirs: @["_test"],
-    skipFiles: @["README.md"],
-    skipExt: @["test"]
-  )
+let opts = Options(sourceDir: sourceDir)
+var pkgInfo = PackageInfo(
+  nimbleFile: sourceDir / "source.nimble",
+  name: "source",
+  version: "1.0",
+  srcDir: "src",
+  skipExt: @["test"]
+)
+
+inferInstallRules(pkgInfo, opts)
 
 let files = pkgInfo.getInstallFiles(opts).sorted
 assert files == @[
-  (pcFile, sourceDir / "source" / "file1.nim"),
-  (pcFile, sourceDir / "source" / "file2.nim"),
-  (pcDir, sourceDir / "source")
+  (pcFile, sourceDir / "src" / "source" / "file1.nim"),
+  (pcFile, sourceDir / "src" / "source" / "file2.nim"),
+  (pcDir, sourceDir / "src" / "source")
 ]